// packages/protocol/typescript/generated/agent_start/runtime/v1/repo_pb.d.ts
export declare type RepoServiceListRequest = Message<"agent_start.runtime.v1.RepoServiceListRequest"> & {
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceListRequest.
 * Use `create(RepoServiceListRequestSchema)` to create a new message.
 */
export declare const RepoServiceListRequestSchema: GenMessage<RepoServiceListRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceListResponse
 */
export declare type RepoServiceListResponse = Message<"agent_start.runtime.v1.RepoServiceListResponse"> & {
  /**
   * @generated from field: repeated agent_start.runtime.v1.Repo repos = 1;
   */
  repos: Repo[];

  /**
   * @generated from field: int64 revision = 2;
   */
  revision: bigint;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceListResponse.
 * Use `create(RepoServiceListResponseSchema)` to create a new message.
 */
export declare const RepoServiceListResponseSchema: GenMessage<RepoServiceListResponse>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceAddRequest
 */
export declare type RepoServiceAddRequest = Message<"agent_start.runtime.v1.RepoServiceAddRequest"> & {
  /**
   * @generated from field: int64 expected_revision = 1;
   */
  expectedRevision: bigint;

  /**
   * @generated from field: string path = 2;
   */
  path: string;

  /**
   * @generated from field: agent_start.runtime.v1.RepoKind kind = 3;
   */
  kind: RepoKind;

  /**
   * @generated from field: optional string host_id = 4;
   */
  hostId?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceAddRequest.
 * Use `create(RepoServiceAddRequestSchema)` to create a new message.
 */
export declare const RepoServiceAddRequestSchema: GenMessage<RepoServiceAddRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceAddResponse
 */
export declare type RepoServiceAddResponse = Message<"agent_start.runtime.v1.RepoServiceAddResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.Repo repo = 1;
   */
  repo?: Repo | undefined;

  /**
   * @generated from field: int64 revision = 2;
   */
  revision: bigint;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceAddResponse.
 * Use `create(RepoServiceAddResponseSchema)` to create a new message.
 */
export declare const RepoServiceAddResponseSchema: GenMessage<RepoServiceAddResponse>;
// ...
export declare type RepoServiceGetHooksRequest = Message<"agent_start.runtime.v1.RepoServiceGetHooksRequest"> & {
  /**
   * @generated from field: string project_id = 1;
   */
  projectId: string;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceGetHooksRequest.
 * Use `create(RepoServiceGetHooksRequestSchema)` to create a new message.
 */
export declare const RepoServiceGetHooksRequestSchema: GenMessage<RepoServiceGetHooksRequest>;
// ...
export declare type RepoServiceGetHooksResponse = Message<"agent_start.runtime.v1.RepoServiceGetHooksResponse"> & {
  /**
   * @generated from field: optional string setup_command = 1;
   */
  setupCommand?: string | undefined;

  /**
   * @generated from field: agent_start.runtime.v1.RepoSetupRunPolicy setup_run_policy = 2;
   */
  setupRunPolicy: RepoSetupRunPolicy;

  /**
   * @generated from field: agent_start.runtime.v1.RepoHooksSource source = 3;
   */
  source: RepoHooksSource;

  /**
   * @generated from field: optional agent_start.runtime.v1.RepoSetupTrust setup_trust = 4;
   */
  setupTrust?: RepoSetupTrust | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceGetHooksResponse.
 * Use `create(RepoServiceGetHooksResponseSchema)` to create a new message.
 */
export declare const RepoServiceGetHooksResponseSchema: GenMessage<RepoServiceGetHooksResponse>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceCloneRequest
 */
export declare type RepoServiceCloneRequest = Message<"agent_start.runtime.v1.RepoServiceCloneRequest"> & {
  /**
   * @generated from field: int64 expected_revision = 1;
   */
  expectedRevision: bigint;

  /**
   * @generated from field: string url = 2;
   */
  url: string;

  /**
   * @generated from field: string destination = 3;
   */
  destination: string;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceCloneRequest.
 * Use `create(RepoServiceCloneRequestSchema)` to create a new message.
 */
export declare const RepoServiceCloneRequestSchema: GenMessage<RepoServiceCloneRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceCloneResponse
 */
export declare type RepoServiceCloneResponse = Message<"agent_start.runtime.v1.RepoServiceCloneResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.Repo repo = 1;
   */
  repo?: Repo | undefined;

  /**
   * @generated from field: int64 revision = 2;
   */
  revision: bigint;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceCloneResponse.
 * Use `create(RepoServiceCloneResponseSchema)` to create a new message.
 */
export declare const RepoServiceCloneResponseSchema: GenMessage<RepoServiceCloneResponse>;
// ...
export declare const RepoServiceCreateRequestSchema: GenMessage<RepoServiceCreateRequest>;
// ...
export declare const RepoServiceCreateResponseSchema: GenMessage<RepoServiceCreateResponse>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceGitAvailableRequest
 */
export declare type RepoServiceGitAvailableRequest = Message<"agent_start.runtime.v1.RepoServiceGitAvailableRequest"> & {
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceGitAvailableRequest.
 * Use `create(RepoServiceGitAvailableRequestSchema)` to create a new message.
 */
export declare const RepoServiceGitAvailableRequestSchema: GenMessage<RepoServiceGitAvailableRequest>;
// ...
export declare const RepoServiceGitAvailableResponseSchema: GenMessage<RepoServiceGitAvailableResponse>;
// ...
export declare const RepoServiceHooksRequestSchema: GenMessage<RepoServiceHooksRequest>;
// ...
export declare const RepoServiceHooksResponseSchema: GenMessage<RepoServiceHooksResponse>;
// ...
export declare const RepoServiceHooksCheckRequestSchema: GenMessage<RepoServiceHooksCheckRequest>;
// ...
export declare const RepoServiceHooksCheckResponseSchema: GenMessage<RepoServiceHooksCheckResponse>;
// ...
export declare const RepoServiceSetupScriptImportsRequestSchema: GenMessage<RepoServiceSetupScriptImportsRequest>;
// ...
export declare const RepoServiceSetupScriptImportsResponseSchema: GenMessage<RepoServiceSetupScriptImportsResponse>;
// ...
export declare type RepoServiceSparsePresetsRequest = Message<"agent_start.runtime.v1.RepoServiceSparsePresetsRequest"> & {
  /**
   * @generated from field: string repo = 1;
   */
  repo: string;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceSparsePresetsRequest.
 * Use `create(RepoServiceSparsePresetsRequestSchema)` to create a new message.
 */
export declare const RepoServiceSparsePresetsRequestSchema: GenMessage<RepoServiceSparsePresetsRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceSparsePresetsResponse
 */
export declare type RepoServiceSparsePresetsResponse = Message<"agent_start.runtime.v1.RepoServiceSparsePresetsResponse"> & {
  /**
   * @generated from field: repeated agent_start.runtime.v1.RepoSparsePreset presets = 1;
   */
  presets: RepoSparsePreset[];
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceSparsePresetsResponse.
 * Use `create(RepoServiceSparsePresetsResponseSchema)` to create a new message.
 */
export declare const RepoServiceSparsePresetsResponseSchema: GenMessage<RepoServiceSparsePresetsResponse>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceSaveSparsePresetRequest
 */
export declare type RepoServiceSaveSparsePresetRequest = Message<"agent_start.runtime.v1.RepoServiceSaveSparsePresetRequest"> & {
  /**
   * @generated from field: string repo = 1;
   */
  repo: string;

  /**
   * @generated from field: optional string id = 2;
   */
  id?: string | undefined;

  /**
   * @generated from field: string name = 3;
   */
  name: string;

  /**
   * @generated from field: repeated string directories = 4;
   */
  directories: string[];
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceSaveSparsePresetRequest.
 * Use `create(RepoServiceSaveSparsePresetRequestSchema)` to create a new message.
 */
export declare const RepoServiceSaveSparsePresetRequestSchema: GenMessage<RepoServiceSaveSparsePresetRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceSaveSparsePresetResponse
 */
export declare type RepoServiceSaveSparsePresetResponse = Message<"agent_start.runtime.v1.RepoServiceSaveSparsePresetResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.RepoSparsePreset preset = 1;
   */
  preset?: RepoSparsePreset | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceSaveSparsePresetResponse.
 * Use `create(RepoServiceSaveSparsePresetResponseSchema)` to create a new message.
 */
export declare const RepoServiceSaveSparsePresetResponseSchema: GenMessage<RepoServiceSaveSparsePresetResponse>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceRemoveSparsePresetRequest
 */
export declare type RepoServiceRemoveSparsePresetRequest = Message<"agent_start.runtime.v1.RepoServiceRemoveSparsePresetRequest"> & {
  /**
   * @generated from field: string repo = 1;
   */
  repo: string;

  /**
   * @generated from field: string preset_id = 2;
   */
  presetId: string;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceRemoveSparsePresetRequest.
 * Use `create(RepoServiceRemoveSparsePresetRequestSchema)` to create a new message.
 */
export declare const RepoServiceRemoveSparsePresetRequestSchema: GenMessage<RepoServiceRemoveSparsePresetRequest>;

/**
 * @generated from message agent_start.runtime.v1.RepoServiceRemoveSparsePresetResponse
 */
export declare type RepoServiceRemoveSparsePresetResponse = Message<"agent_start.runtime.v1.RepoServiceRemoveSparsePresetResponse"> & {
  /**
   * @generated from field: bool removed = 1;
   */
  removed: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.RepoServiceRemoveSparsePresetResponse.
 * Use `create(RepoServiceRemoveSparsePresetResponseSchema)` to create a new message.
 */
export declare const RepoServiceRemoveSparsePresetResponseSchema: GenMessage<RepoServiceRemoveSparsePresetResponse>;
// ...
export declare const RepoServiceReorderRequestSchema: GenMessage<RepoServiceReorderRequest>;
// ...
export declare const RepoServiceReorderResponseSchema: GenMessage<RepoServiceReorderResponse>;
// ...
export declare const RepoServiceRmRequestSchema: GenMessage<RepoServiceRmRequest>;
// ...
export declare const RepoServiceRmResponseSchema: GenMessage<RepoServiceRmResponse>;
// ...
export declare const RepoServiceUpdateRequestSchema: GenMessage<RepoServiceUpdateRequest>;
// ...
export declare const RepoServiceUpdateResponseSchema: GenMessage<RepoServiceUpdateResponse>;
// ...
export declare const RepoServiceBaseRefDefaultRequestSchema: GenMessage<RepoServiceBaseRefDefaultRequest>;
// ...
export declare const RepoServiceBaseRefDefaultResponseSchema: GenMessage<RepoServiceBaseRefDefaultResponse>;
// ...
export declare const RepoServiceSearchRefsRequestSchema: GenMessage<RepoServiceSearchRefsRequest>;
// ...
export declare const RepoServiceSearchRefsResponseSchema: GenMessage<RepoServiceSearchRefsResponse>;
// ...
export declare const RepoService: GenService<{
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.List
   */
  list: {
    methodKind: "unary";
    input: typeof RepoServiceListRequestSchema;
    output: typeof RepoServiceListResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Add
   */
  add: {
    methodKind: "unary";
    input: typeof RepoServiceAddRequestSchema;
    output: typeof RepoServiceAddResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.GetHooks
   */
  getHooks: {
    methodKind: "unary";
    input: typeof RepoServiceGetHooksRequestSchema;
    output: typeof RepoServiceGetHooksResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Clone
   */
  clone: {
    methodKind: "unary";
    input: typeof RepoServiceCloneRequestSchema;
    output: typeof RepoServiceCloneResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Create
   */
  create: {
    methodKind: "unary";
    input: typeof RepoServiceCreateRequestSchema;
    output: typeof RepoServiceCreateResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.GitAvailable
   */
  gitAvailable: {
    methodKind: "unary";
    input: typeof RepoServiceGitAvailableRequestSchema;
    output: typeof RepoServiceGitAvailableResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Hooks
   */
  hooks: {
    methodKind: "unary";
    input: typeof RepoServiceHooksRequestSchema;
    output: typeof RepoServiceHooksResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.HooksCheck
   */
  hooksCheck: {
    methodKind: "unary";
    input: typeof RepoServiceHooksCheckRequestSchema;
    output: typeof RepoServiceHooksCheckResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.SetupScriptImports
   */
  setupScriptImports: {
    methodKind: "unary";
    input: typeof RepoServiceSetupScriptImportsRequestSchema;
    output: typeof RepoServiceSetupScriptImportsResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.SparsePresets
   */
  sparsePresets: {
    methodKind: "unary";
    input: typeof RepoServiceSparsePresetsRequestSchema;
    output: typeof RepoServiceSparsePresetsResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.SaveSparsePreset
   */
  saveSparsePreset: {
    methodKind: "unary";
    input: typeof RepoServiceSaveSparsePresetRequestSchema;
    output: typeof RepoServiceSaveSparsePresetResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.RemoveSparsePreset
   */
  removeSparsePreset: {
    methodKind: "unary";
    input: typeof RepoServiceRemoveSparsePresetRequestSchema;
    output: typeof RepoServiceRemoveSparsePresetResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Reorder
   */
  reorder: {
    methodKind: "unary";
    input: typeof RepoServiceReorderRequestSchema;
    output: typeof RepoServiceReorderResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Rm
   */
  rm: {
    methodKind: "unary";
    input: typeof RepoServiceRmRequestSchema;
    output: typeof RepoServiceRmResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.Update
   */
  update: {
    methodKind: "unary";
    input: typeof RepoServiceUpdateRequestSchema;
    output: typeof RepoServiceUpdateResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.BaseRefDefault
   */
  baseRefDefault: {
    methodKind: "unary";
    input: typeof RepoServiceBaseRefDefaultRequestSchema;
    output: typeof RepoServiceBaseRefDefaultResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.RepoService.SearchRefs
   */
  searchRefs: {
    methodKind: "unary";
    input: typeof RepoServiceSearchRefsRequestSchema;
    output: typeof RepoServiceSearchRefsResponseSchema;
  },
}>;